refactor-#158: move booking resolvers to folders
