+++ src/microservices/booking/graphql/mutations/index.ts
@@ -1,0 +1,1 @@
+export * from './booking-mutation.resolver';
