+++ src/microservices/booking/graphql/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './booking-query.resolver';
